refactor: share one settings UI between the popup and the sidebar
The sidebar was a 1468-line near-copy of the popup's 1420 lines — the same
screen in a different frame, kept in sync by hand. Diffing them showed 74
differing lines, nearly all comments. Every setting added recently was written
twice and verified once, which is exactly the drift the sidebar parity test
was added to catch.

The behaviour now lives in ui/settingsUi.ts, imported by both entrypoints,
which are two lines each.

The one real difference was rules import: the popup opens a dedicated page
(deliberately — Firefox closes the popup when a file dialog takes focus) while
the sidebar had an inline file picker. Both now use the dedicated page, which
has neither problem, so the sidebar's picker and the dead file input in both
documents are gone.

Markup is still duplicated across the two documents; the parity test covers
that. Sharing it too would mean templating the HTML, which is a bigger change
than this.
