+++ entrypoints/popup/index.html
@@ -283,7 +283,6 @@
               >
                 Import
               </button>
-              <input type="file" id="importFileInput" accept=".json" style="display: none" />
             </div>
             <div class="help-text export-import-help" data-i18n="rulesExportImportHelp">
               Export/import rules as JSON files for backup or sharing
