+++ entrypoints/sidebar/index.html
@@ -284,7 +284,6 @@
                 >
                   Import
                 </button>
-                <input type="file" id="importFileInput" accept=".json" style="display: none" />
               </div>
               <div class="help-text export-import-help" data-i18n="rulesExportImportHelp">
                 Export/import rules as JSON files for backup or sharing
